python311Packages.scikit-build: suppress chmod errors
The scikit-build package in its setuptools wrapper tries to chmod certain
executables it uses to build. These executables are, in our case, in the
immutable nix store and already come equipped with a reasonable set of
permissions. So don't fail outright, but suppress the inevtiable
PermissionError, so builds can continue regardless.